requirements.txt
pyrealsense2
tensorflow
numpy
opencv-python
pyyaml
python3-rospy

dlib
